Fix the leipzig parser
The Studentenwerk Leipzig recently launched a new website that broke the old
parser, which relied on an ajax api that was used in the old website. The new
one offers a dedicated xml API [1].

The new parser uses BeautifulSoup to parse the xml and therefore needs lxml.

It also fixes the problem of the previous parser which added every component
of a meal as a separate meal.

[1]: https://www.studentenwerk-leipzig.de/XMLInterface/