karlsruhe: parse notes and closed days (#95)
The meal names contained the notes in square brackets. This commit properly
parses them and outputs them as notes. It also considers the little icons
in front of the meals. The mapping consists of a hard-coded part copied from
the official additives PDF and a dynamic part from the website. The hard-coded
mapping is needed as the one on the website is not complete.

If a meal name like "Geschlossen von: Montag 18.02. - Freitag 29.03." is
found, the corresponding dates are marked as closed.