ostniedersachsen: Rewrite the parser based on the API
The new API [1] returns a simple xml file that contains all meals of all
canteens. To avoid stressing the API too much, we cache the xml file for
15 minutes.

The xml format also contains the name and address of the canteens, which
we can use for the metadata feed.

Unfortunately, it doesn't contain a description of the additives and
allergenes currently, so have put this directly into the parser for now.

Fix #99.